refactor(ollama): extract shared functions into adapter_shared.rs
